refactor configs based on feedback
